+++ lib/src/http_request_task.dart
@@ -11,9 +11,12 @@ class HttpRequestTask {
   final Encoding? encoding;
   final Map<String, String>? formFields;
   final bool silent;
+  final bool authenticate;
   final int? timeoutSecs;
   final String? requestId;
+  final String? savePath;
   final Completer<dynamic> completer;
+  final void Function(int received, int total)? onReceiveProgress;
 
   HttpRequestTask(this.completer, this.method, this.uu, this.headers,
       {this.params,
@@ -21,5 +24,8 @@ class HttpRequestTask {
       this.formFields,
       this.timeoutSecs,
       this.requestId,
+      this.onReceiveProgress,
+      this.savePath = "",
+      this.authenticate = true,
       this.silent = false});
 }